Fixing login validation
