+++ presentation/src/main/java/com/bladoae/imdb/presentation/common/ErrorMessage.kt
@@ -1,0 +1,56 @@
+package com.bladoae.imdb.presentation.common
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Button
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.bladoae.imdb.presentation.R
+import com.bladoae.imdb.presentation.theme.IMDbTheme
+
+@Composable
+fun ErrorMessage(
+    onRetry: () -> Unit
+) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(24.dp),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Text(
+            modifier = Modifier.padding(8.dp),
+            text = stringResource(id = R.string.error_message)
+        )
+        Button(
+            onClick = { onRetry() },
+            content = {
+                Text(
+                    text = stringResource(id = R.string.retry),
+                    style = MaterialTheme.typography.body1,
+                )
+            }
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun ErrorMessagePreview() {
+    IMDbTheme {
+        ErrorMessage(
+            {
+
+            }
+        )
+    }
+}
