+++ presentation/src/main/java/com/bladoae/imdb/presentation/common/LoadingMessage.kt
@@ -1,0 +1,30 @@
+package com.bladoae.imdb.presentation.common
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import com.bladoae.imdb.presentation.theme.IMDbTheme
+
+@Composable
+fun LoadingMessage() {
+    Row(
+        modifier = Modifier.fillMaxSize(),
+        horizontalArrangement = Arrangement.Center,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        CircularProgressIndicator()
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun LoadingMessagePreview() {
+    IMDbTheme {
+        LoadingMessage()
+    }
+}
